requirements.txt
# Основные зависимости проекта (без cog-whisper-diarization)
torch
torchaudio
pyannote.audio
